Barebone views for display
- Added bare bone views for display
- Refactored the display page to render a view based on state
- New methods that allow subscribing (and unsubscribing) to the command stream
- Remove socket related logic from the game view and only use the new subscribing methods
