+++ helpers.js
@@ -1,0 +1,4 @@
+export const isBrowser = () => {
+  if (typeof window != "undefined") return true;
+  return false;
+}
