+++ lib/views.js
@@ -1,1 +1,4 @@
 export const DISPLAY_WAITING = 'DISPLAY_WAITING'
+export const DISPLAY_GAME = 'DISPLAY_GAME'
+export const DISPLAY_WAITING_TO_START = 'DISPLAY_WAITING_TO_START'
+export const DISPALY_GAME_OVER = 'DISPALY_GAME_OVER'
